third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/security/suborigins/crossorigin/suborigin-cors-script-expected.txt

CONSOLE ERROR: Access to Script at 'http://127.0.0.1:8000/security/resources/cors-script.php?value=ran' from origin 'http-so://foobar.127.0.0.1:8000' has been blocked by CORS policy: The 'Access-Control-Allow-Origin' header has a value 'http://127.0.0.1:8000' that is not equal to the supplied origin. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Script at 'http://127.0.0.1:8000/security/resources/cors-script.php?credentials=true&value=ran' from origin 'http-so://foobar.127.0.0.1:8000' has been blocked by CORS policy: The 'Access-Control-Allow-Origin' header has a value 'http://127.0.0.1:8000' that is not equal to the supplied origin. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Script at 'http://127.0.0.1:8000/security/resources/cors-script.php?value=ran&cors=false' from origin 'http-so://foobar.127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http-so://foobar.127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS Script: <crossorigin='anonymous'>, ACAO: http://127.0.0.1:8000 
PASS Script: <crossorigin='anonymous'>, ACAO: * 
PASS Script: <crossorigin='use-credentials'>, ACAO: http://127.0.0.1:8000 
PASS Script: <crossorigin='anonymous'>, CORS-ineligible resource 
Harness: the test ran to completion.
